refactor: EditProfileModalをpagesレイヤーに移行
- EditProfileModalをsrc/pages/profile/uiに移動
- "use client"ディレクティブを追加
- Storybookファイルもプロフィールページに移動
- shared/uiからEditProfileModalのエクスポートを削除
- ProfilePageClientでローカルインポートに変更
- FSD準拠性を確保
