+++ plan.md
@@ -4,3 +4,4 @@
 - [x] `src/shared/ui/Post/Post.tsx` を `src/entities/post` へ移動する
 ✅ `src/shared/ui/PostActionMenu` を `src/entities/post` へ移動する
 ✅ entities や features などの ui フォルダ直下の index.ts を削除する。
+✅ `src/shared/ui` 内のファイルで特定のページからしか使用されていないコンポーネントを `src/pages/〇〇/ui` へ移動する
