+++ src/pages/profile/index.ts
@@ -3,3 +3,4 @@ export { Profile } from "./ui/Profile";
 export { ProfileHeader } from "./ui/ProfileHeader";
 export { ProfileActionsButton } from "./ui/ProfileActionsButton";
 export { ProfileStats } from "./ui/ProfileStats";
+export { EditProfileModal } from "./ui/EditProfileModal";
