+++ src/pages/profile/ui/ProfilePageClient.tsx
@@ -2,9 +2,10 @@
 
 import type { PostType, ProfileType, User } from "@/shared/entities";
 import { sideBarWidthBase } from "@/shared/lib/const";
-import { EditProfileModal, Sidebar } from "@/shared/ui";
+import { Sidebar } from "@/shared/ui";
 import { PostsList } from "@/widgets/posts-list";
 import { Box, Flex } from "@mantine/core";
+import { EditProfileModal } from "./EditProfileModal";
 
 import { Profile } from "./Profile";
