+++ src/shared/ui/index.ts
@@ -1,6 +1,5 @@
 export { Sidebar } from "./Sidebar/Sidebar";
 export { ImageUpload } from "./ImageUpload/ImageUpload";
-export { EditProfileModal } from "./Modal/EditProfileModal/EditProfileModal";
 export { NewPostForm } from "./NewPostForm/NewPostForm";
 export { CenteredLoader } from "./CenteredLoader/CenteredLoader";
 export { ContentPart, parseContent } from "./ContentPart/ContentPart";
